requirements.txt
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
lxml==5.2.1
magic-filter==1.0.12
multidict==6.0.5
outcome==1.3.0.post0
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.1.0
pyppeteer==2.0.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.20.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==1.26.18
websockets==10.4
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
